+++ tutu-frontend/src/lib/localization/messages/fi.json
@@ -274,8 +274,8 @@
   },
   "maajako": {
     "otsikko": "Esittelijöiden maajako",
-    "otsikko2": "Tarkista kenelle maa on valittu",
-    "otsikko3": "Maajako",
+    "kenellevalittu": "Tarkista kenelle maa on valittu",
+    "maajako": "Maajako",
     "kuvaus": "Hallinnoi maajakoa, jonka perusteella hakemukset ohjautuvat oikealle esittelijälle. Kukin maa voi olla valittuna vain yhdelle esittelijälle. Hakemukset ohjautuvat sen perusteella, minkä maan hakija on ilmoittanut hakemuslomakkeella tutkinnon suoritusmaaksi kohdassa Tutkinto 1.",
     "suoritusmaa": "Tutkinnon suoritusmaa",
     "kaikkiValittu": "Kaikki suoritusmaat on valittu jollekin esittelijälle",
